pyproject.toml
[tool.poetry]
name = "whisper-hailo-8l-fastapi"
version = "0.1.0"
description = ""
authors = ["MafiaCoconut <alex.kach.2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
scipy = "^1.16.0"
torch = "^2.7.1"
sounddevice = "^0.5.2"
transformers = "^4.52.4"
uvicorn = "^0.34.3"
unicorn = "^2.1.3"
fastapi = "^0.115.13"
icecream = "^2.1.4"
colorlog = "^6.9.0"
coloredlogs = "^15.0.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
